requirements.txt
PySide6==6.3.1
requests
PyGithub